openapiart/requirements.txt
--prefer-binary
jsonpath-ng
typing
setuptools
wheel
PyYAML
openapi_spec_validator==0.3.0
requests
typing_extensions ; python_version > '2.7'
click==8.0.4; python_version > '2.7' # this is added because of the black failure in CI
black==22.1.0 ; python_version > '2.7'
grpcio-tools~=1.54.2 ; python_version > '3.6'
grpcio-tools~=1.35.0 ; python_version <= '3.6'
pipreqs==0.4.11; python_version > '2.7'
protobuf~=3.15.0 ;  python_version <= '3.6'
protobuf~=4.23.3 ;  python_version > '3.6'
grpcio~=1.54.2 ; python_version > '3.6'
grpcio~=1.35.0 ; python_version <= '3.6'
